new file for plot example
