+++ ACPerfSol/mainwindow.cpp
@@ -19,46 +19,6 @@ MainWindow::MainWindow(QWidget *parent) :
     ui->customPlot->replot();
 }
 
-void MainWindow::setupLineStyleDemo(QCustomPlot *customPlot)
-{
-  demoName = "Line Style Demo";
-  customPlot->legend->setVisible(true);
-  customPlot->legend->setFont(QFont("Helvetica", 9));
-  QPen pen;
-  QStringList lineNames;
-  lineNames << "lsNone" << "lsLine" << "lsStepLeft" << "lsStepRight" << "lsStepCenter" << "lsImpulse";
-  // add graphs with different line styles:
-  for (int i=QCPGraph::lsNone; i<=QCPGraph::lsImpulse; ++i)
-  {
-    customPlot->addGraph();
-    pen.setColor(QColor(qSin(i*1+1.2)*80+80, qSin(i*0.3+0)*80+80, qSin(i*0.3+1.5)*80+80));
-    customPlot->graph()->setPen(pen);
-    customPlot->graph()->setName(lineNames.at(i-QCPGraph::lsNone));
-    customPlot->graph()->setLineStyle((QCPGraph::LineStyle)i);
-    customPlot->graph()->setScatterStyle(QCPScatterStyle(QCPScatterStyle::ssCircle, 5));
-    // generate data:
-    QVector<double> x(15), y(15);
-    for (int j=0; j<15; ++j)
-    {
-      x[j] = j/15.0 * 5*3.14 + 0.01;
-      y[j] = 7*qSin(x[j])/x[j] - (i-QCPGraph::lsNone)*5 + (QCPGraph::lsImpulse)*5 + 2;
-    }
-    customPlot->graph()->setData(x, y);
-    customPlot->graph()->rescaleAxes(true);
-  }
-  // zoom out a bit:
-  customPlot->yAxis->scaleRange(1.1, customPlot->yAxis->range().center());
-  customPlot->xAxis->scaleRange(1.1, customPlot->xAxis->range().center());
-  // set blank axis lines:
-  customPlot->xAxis->setTicks(false);
-  customPlot->yAxis->setTicks(true);
-  customPlot->xAxis->setTickLabels(false);
-  customPlot->yAxis->setTickLabels(true);
-  // make top right axes clones of bottom left axes:
-  customPlot->axisRect()->setupFullAxesBox();
-}
-
-
 MainWindow::~MainWindow()
 {
     delete ui;
